update file and collections to modules
